requirements.txt
pylint
pytest
Flask==1.1.1
gunicorn
google-cloud-bigquery
pandas
pandas-gbq
google-api-python-client
google-cloud
google-cloud-ndb
google-cloud-datastore